task-manager/requirements.txt
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
